Refactor and Testing
Renames anything called 'ConductionMatrix' to 'ConductanceMatrix' and includes some test code in Source.cpp
