--- spice-simulator/ComplexMatrix.h
@@ -1,9 +1,0 @@
-#pragma once
-
-#include <complex>
-#include <Eigen/Dense>
-#include "Component.h"
-
-namespace Eigen {
-	typedef Matrix<std::complex<double>, Dynamic, 1> ComplexVector;
-}
