+++ spice-simulator/ConductanceMatrix.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <complex>
+#include <Eigen/Dense>
+#include "Component.h"
+
+Eigen::MatrixXcd getConductanceMatrix(std::vector<Component*> components, int numNodes);
