+++ spice-simulator/Source.cpp
@@ -1,14 +1,17 @@
 #include <iostream>
+#include "ConductanceMatrix.h"
 #include "Component.h"
 
 int main() {
-
 	std::vector<Component*> components;
 
-	Resistor r("R1", 100, 0, 1);
+	CurrentSource i1("I1", 1, 0, 1);
+	Resistor r1("R1", 100, 1, 2);
+	Resistor r2("R2", 200, 2, 0);
 
-	components.push_back(&r);
+	components.push_back(&i1);
+	components.push_back(&r1);
+	components.push_back(&r2);
 
-	std::cout << components[0]->getConductance(0, 1) << std::endl;
-
+	std::cout << getConductanceMatrix(components, 2) << std::endl;
 }
