+++ spice-simulator/spice-simulator.vcxproj
@@ -143,11 +143,11 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="Component.cpp" />
-    <ClCompile Include="ConductionMatrix.cpp" />
+    <ClCompile Include="ConductanceMatrix.cpp" />
     <ClCompile Include="Source.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="ComplexMatrix.h" />
+    <ClInclude Include="ConductanceMatrix.h" />
     <ClInclude Include="Component.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
